# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple blog API built with Go and Gin",
        "title": "Blog API",
        "contact": {
            "name": "Michael Obeng",
            "url": "https://github.com/pandahawk",
            "email": "michael@example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
      "/posts": {
        "get": {
          "description": "Get all posts in the system",
          "produces": [
            "application/json"
          ],
          "tags": [
            "posts"
          ],
          "summary": "Get all posts",
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/model.Post"
                }
              }
            }
          }
        },
        "post": {
          "description": "Creates a new post and returns the created resource",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "posts"
          ],
          "summary": "Create a new post",
          "parameters": [
            {
              "description": "Post data",
              "name": "post",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/post.CreatePostRequest"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Created",
              "schema": {
                "$ref": "#/definitions/post.Response"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/apperrors.InvalidInputError"
              }
            },
            "409": {
              "description": "Conflict",
              "schema": {
                "$ref": "#/definitions/apperrors.DuplicateError"
              }
            }
          }
        }
      },
      "/posts/{id}": {
        "get": {
          "description": "Get the post with the specified ID",
          "produces": [
            "application/json"
          ],
          "tags": [
            "posts"
          ],
          "summary": "Get post by ID",
          "parameters": [
            {
              "type": "string",
              "description": "Post ID",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/post.Response"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/apperrors.InvalidInputError"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/apperrors.NotFoundError"
              }
            }
          }
        },
        "delete": {
          "description": "Deletes an existing post",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "posts"
          ],
          "summary": "Delete post by ID",
          "parameters": [
            {
              "type": "string",
              "format": "uuid",
              "description": "Post ID",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "204": {
              "description": "No Content"
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/apperrors.InvalidInputError"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/apperrors.NotFoundError"
              }
            }
          }
        },
        "patch": {
          "description": "Updates an existing post and returns the updated resource",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "posts"
          ],
          "summary": "Update post by ID",
          "parameters": [
            {
              "type": "string",
              "format": "uuid",
              "description": "Post ID",
              "name": "id",
              "in": "path",
              "required": true
            },
            {
              "description": "Post update data",
              "name": "post",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/post.UpdatePostRequest"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Created",
              "schema": {
                "$ref": "#/definitions/post.Response"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/apperrors.DuplicateError"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/apperrors.NotFoundError"
              }
            }
          }
        }
      },
        "/users": {
            "get": {
                "description": "Get all users in the system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                              "$ref": "#/definitions/user.Response"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new user and returns the created resource",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "User data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                          "$ref": "#/definitions/user.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                          "$ref": "#/definitions/user.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.InvalidInputError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/apperrors.DuplicateError"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Get the user with the specified ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                          "$ref": "#/definitions/user.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.InvalidInputError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apperrors.NotFoundError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an existing user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Delete user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.InvalidInputError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apperrors.NotFoundError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates an existing user and returns the updated resource",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User update data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                          "$ref": "#/definitions/user.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                          "$ref": "#/definitions/user.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.DuplicateError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apperrors.NotFoundError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apperrors.DuplicateError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                }
            }
        },
        "apperrors.InvalidInputError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "apperrors.NotFoundError": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "resource": {
                    "type": "string"
                }
            }
        },
      "model.Post": {
            "type": "object",
            "required": [
                "content",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "example": "My First Post Content"
                },
              "createdAt": {
                    "type": "string",
                    "example": "2025-07-18T15:04:05Z"
                },
                "id": {
                    "type": "string",
                    "example": "4e76b320-d5b7-4a0a-bb0f-2049fe6a91a7"
                },
                "title": {
                    "type": "string",
                    "example": "My First Post"
                },
              "updatedAt": {
                    "type": "string",
                    "example": "2025-08-19T15:04:05Z"
                },
              "user": {
                "$ref": "#/definitions/model.User"
              },
                "userID": {
                  "type": "string"
                }
            }
        },
      "model.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-07-18T15:04:05Z"
                },
                "email": {
                    "type": "string",
                    "example": "mike@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "b9e69a63-4f4b-4ea7-8c71-3b73fe62e6d7"
                },
                "posts": {
                    "type": "array",
                    "items": {
                      "$ref": "#/definitions/model.Post"
                    }
                },
                "username": {
                    "type": "string",
                    "example": "mike"
                }
            }
      },
      "post.CreatePostRequest": {
        "type": "object",
        "required": [
          "author_id",
          "content",
          "title"
        ],
        "properties": {
          "author_id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "post.Response": {
        "type": "object",
        "properties": {
          "author": {
            "$ref": "#/definitions/post.UserSummaryResponse"
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "post_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
      "post.UpdatePostRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "post.UserSummaryResponse": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "user.CreateUserRequest": {
        "type": "object",
        "required": [
          "email",
          "username"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "user.PostSummaryResponse": {
        "type": "object",
        "properties": {
          "post_id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "example": "My First Post"
          }
        }
      },
      "user.Response": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "joined_at": {
            "type": "string"
          },
          "posts": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/user.PostSummaryResponse"
            }
          },
          "user_id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "user.UpdateUserRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
        }
    }
}